moved upload file from pageObject to separate file
